1SmartShopperAi(Ali-Ayman)/requirements.txt
crewai==0.126.0
crewai-tools==0.47.0
pydantic==2.11.5
scrapegraph-py==1.12.0
tavily-python==0.7.4
python-dotenv==1.1.0
streamlit==1.45.0
pysqlite3-binary
